Moved AudioTimePipe to its own file
And removed duplicate
