+++ src/SIL.XForge.Scripture/ClientApp/src/app/checking/checking/checking-audio-player-new/checking-audio-player-new.component.spec.ts
@@ -9,7 +9,7 @@ import { PwaService } from 'xforge-common/pwa.service';
 import { TestTranslocoModule } from 'xforge-common/test-utils';
 import { UICommonModule } from 'xforge-common/ui-common.module';
 import { AudioPlayerComponent } from '../../../shared/audio/audio-player.component/audio-player.component';
-import { AudioTimePipe } from '../checking-audio-player/checking-audio-player.component';
+import { AudioTimePipe } from '../../../shared/audio/audio-time-pipe';
 import { CheckingAudioPlayerNewComponent } from './checking-audio-player-new.component';
 
 describe('CheckingAudioPlayerNewComponent', () => {
