+++ src/SIL.XForge.Scripture/ClientApp/src/app/checking/checking/checking-audio-player/checking-audio-player.component.spec.ts
@@ -8,7 +8,8 @@ import { I18nService } from 'xforge-common/i18n.service';
 import { PwaService } from 'xforge-common/pwa.service';
 import { TestTranslocoModule } from 'xforge-common/test-utils';
 import { UICommonModule } from 'xforge-common/ui-common.module';
-import { AudioStatus, AudioTimePipe, CheckingAudioPlayerComponent } from './checking-audio-player.component';
+import { AudioTimePipe } from '../../../shared/audio/audio-time-pipe';
+import { AudioStatus, CheckingAudioPlayerComponent } from './checking-audio-player.component';
 
 describe('CheckingAudioPlayerComponent', () => {
   const audioFile = 'test-audio-player.webm';
