+++ src/SIL.XForge.Scripture/ClientApp/src/app/shared/audio/audio-time-pipe.ts
@@ -1,0 +1,11 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({ name: 'audioTime' })
+export class AudioTimePipe implements PipeTransform {
+  transform(seconds: number, ..._args: any[]): string {
+    const minutesString = Math.floor(seconds / 60);
+    seconds = Math.floor(seconds % 60);
+    const secondsString = seconds >= 10 ? seconds : '0' + seconds;
+    return minutesString + ':' + secondsString;
+  }
+}
